Simplify structure Fix namespaces
