+++ www/tests/EnvTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests;
+namespace App\Tests;
 
 /**
  * Class EnvTest
