+++ www/tests/helpers/WebTestCase.php
@@ -1,19 +1,20 @@
 <?php
 
-namespace Tests\Helpers;
+namespace App\Tests\Helpers;
 
-use Tests\Helpers\WebTestClient;
+/**
+ * Class LocalWebTestCase
+ * @package Tests
+ *
+ * Class for performing application testing with HTTP Requests
+ */
+class WebTestCase extends \PHPUnit\Framework\TestCase {
 
-abstract class WebTestCase extends \PHPUnit\Framework\TestCase
-{
     /** @var \Slim\App */
     protected $app;
     /** @var WebTestClient */
     protected $client;
 
-    /**
-     * Run for each unit test to setup our slim app environment
-     */
     public function setup()
     {
         // Establish a local reference to the Slim app object
@@ -23,7 +24,33 @@ abstract class WebTestCase extends \PHPUnit\Framework\TestCase
     }
 
     /**
-     * Abstract method to get slim instance
+     * Register application instance for testing
      */
-    abstract protected function getSlimInstance();
-}
+    public function getSlimInstance() {
+        $app = new \Slim\App([
+            'settings' => ['displayErrorDetails' => false]
+        ]);
+
+        /**
+         * Load Dotenv
+         * @see https://github.com/vlucas/phpdotenv
+         */
+        if (is_file(PROJECTROOT . '.env')) {
+            $dotenv = new \Dotenv\Dotenv(PROJECTROOT);
+            $dotenv->load();
+        }
+
+        /**
+         * Enable modules
+         */
+        include PROJECTROOT . 'public/modules.php';
+
+        /**
+         * Set routes
+         */
+        include PROJECTROOT . 'public/routes.php';
+
+        return $app;
+    }
+};
+
